Remove ban, replace with timeout
These scams spread like worm viruses. They scam users, and the scammed users scam other users in friends list and guilds.

The scammed users, in many cases, do get their accounts back. And they also find they were banned on many of their beloved guilds.

This happened to two of my friends. No hacker will hang on dearly to one random account for *1 year*. Most of the time it's just for selling tokens.

The user will just contact an admin, and ask to be unbanned. The admin will tell the user to change their password to reset the token, and no hacker mass spamming guilds and friends list will pretend to be a user to spam again on a server with *automatic moderation*.